move processing options into relevant partial classes
