+++ WarpLib/TiltSeries/TiltSeries.cs
@@ -2504,22 +2504,7 @@ namespace Warp
         }
     }
 
-    [Serializable]
-    public class ProcessingOptionsTomoStack : TomoProcessingOptionsBase
-    {
-        [WarpSerializable]
-        public bool ApplyMask { get; set; }
-    }
 
-    [Serializable]
-    public class ProcessingOptionsTomoImportAlignments : TomoProcessingOptionsBase
-    {
-        [WarpSerializable]
-        public decimal MinFOV { get; set; }
-
-        [WarpSerializable]
-        public string OverrideResultsDir { get; set; }
-    }
 
     [Serializable]
     public class ProcessingOptionsTomoAretomo : TomoProcessingOptionsBase
@@ -2585,172 +2570,5 @@ namespace Warp
         
         [WarpSerializable]
         public decimal TargetNBeads { get; set; }
-    }
-
-
-    [Serializable]
-    public class ProcessingOptionsTomoFullReconstruction : TomoProcessingOptionsBase
-    {
-        [WarpSerializable]
-        public bool OverwriteFiles { get; set; }
-
-        [WarpSerializable]
-        public bool Invert { get; set; }
-
-        [WarpSerializable]
-        public bool Normalize { get; set; }
-
-        [WarpSerializable]
-        public bool DoDeconv { get; set; }
-
-        [WarpSerializable]
-        public decimal DeconvStrength { get; set; }
-
-        [WarpSerializable]
-        public decimal DeconvFalloff { get; set; }
-
-        [WarpSerializable]
-        public decimal DeconvHighpass { get; set; }
-
-        [WarpSerializable]
-        public int SubVolumeSize { get; set; }
-
-        [WarpSerializable]
-        public decimal SubVolumePadding { get; set; }
-
-        [WarpSerializable]
-        public bool PrepareDenoising { get; set; }
-
-        [WarpSerializable]
-        public bool PrepareDenoisingFrames { get; set; }
-
-        [WarpSerializable]
-        public bool PrepareDenoisingTilts { get; set; }
-
-        [WarpSerializable]
-        public bool KeepOnlyFullVoxels { get; set; }
-
-        public override bool Equals(object obj)
-        {
-            if (ReferenceEquals(null, obj)) return false;
-            if (ReferenceEquals(this, obj)) return true;
-            if (obj.GetType() != this.GetType()) return false;
-            return Equals((ProcessingOptionsTomoFullReconstruction)obj);
-        }
-
-        protected bool Equals(ProcessingOptionsTomoFullReconstruction other)
-        {
-            return base.Equals(other) &&
-                   Invert == other.Invert &&
-                   Normalize == other.Normalize &&
-                   DoDeconv == other.DoDeconv &&
-                   DeconvStrength == other.DeconvStrength &&
-                   DeconvFalloff == other.DeconvFalloff &&
-                   DeconvHighpass == other.DeconvHighpass &&
-                   SubVolumeSize == other.SubVolumeSize &&
-                   SubVolumePadding == other.SubVolumePadding &&
-                   PrepareDenoising == other.PrepareDenoising &&
-                   PrepareDenoisingFrames == other.PrepareDenoisingFrames &&
-                   PrepareDenoisingTilts == other.PrepareDenoisingTilts &&
-                   KeepOnlyFullVoxels == other.KeepOnlyFullVoxels;
-        }
-
-        public static bool operator ==(ProcessingOptionsTomoFullReconstruction left, ProcessingOptionsTomoFullReconstruction right)
-        {
-            return Equals(left, right);
-        }
-
-        public static bool operator !=(ProcessingOptionsTomoFullReconstruction left, ProcessingOptionsTomoFullReconstruction right)
-        {
-            return !Equals(left, right);
-        }
-    }
-
-    [Serializable]
-    public class ProcessingOptionsTomoFullMatch : TomoProcessingOptionsBase
-    {
-        [WarpSerializable]
-        public bool OverwriteFiles { get; set; }
-
-        [WarpSerializable]
-        public int SubVolumeSize { get; set; }
-
-        [WarpSerializable]
-        public string TemplateName { get; set; }
-
-        [WarpSerializable]
-        public decimal TemplatePixel { get; set; }
-
-        [WarpSerializable]
-        public decimal TemplateDiameter { get; set; }
-
-        [WarpSerializable]
-        public decimal PeakDistance { get; set; }
-
-        [WarpSerializable]
-        public decimal TemplateFraction { get; set; }
-
-        [WarpSerializable]
-        public int MaxMissingTilts { get; set; }
-
-        [WarpSerializable]
-        public bool WhitenSpectrum { get; set; }
-
-        [WarpSerializable]
-        public decimal Lowpass { get; set; }
-
-        [WarpSerializable]
-        public decimal LowpassSigma { get; set; }
-
-        [WarpSerializable]
-        public string Symmetry { get; set; }
-
-        [WarpSerializable]
-        public int HealpixOrder { get; set; }
-
-        [WarpSerializable]
-        public decimal TiltRange { get; set; }
-
-        [WarpSerializable]
-        public int BatchAngles { get; set; }
-
-        [WarpSerializable]
-        public int Supersample { get; set; }
-
-        [WarpSerializable]
-        public int NResults { get; set; }
-
-        [WarpSerializable]
-        public bool NormalizeScores { get; set; }
-
-        [WarpSerializable]
-        public bool ReuseCorrVolumes { get; set; }
-    }
-
-    [Serializable]
-    public class ProcessingOptionsTomoSubReconstruction : TomoProcessingOptionsBase
-    {
-        [WarpSerializable]
-        public string Suffix { get; set; }
-        [WarpSerializable]
-        public int BoxSize { get; set; }
-        [WarpSerializable]
-        public int ParticleDiameter { get; set; }
-        [WarpSerializable]
-        public bool Invert { get; set; }
-        [WarpSerializable]
-        public bool NormalizeInput { get; set; }
-        [WarpSerializable]
-        public bool NormalizeOutput { get; set; }
-        [WarpSerializable]
-        public bool PrerotateParticles { get; set; }
-        [WarpSerializable]
-        public bool DoLimitDose { get; set; }
-        [WarpSerializable]
-        public int NTilts { get; set; }
-        [WarpSerializable]
-        public bool MakeSparse { get; set; }
-        [WarpSerializable]
-        public bool UseCPU { get; set; }
     }
 }
